% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidTimeSeries.R
\name{covidTimeSeries}
\alias{covidTimeSeries}
\title{Plot covid-19 time series}
\usage{
covidTimeSeries(
  country = "egypt",
  min.date = "2020-01-22",
  max.date = as.Date(today())
)
}
\arguments{
\item{country}{name of the country}

\item{min.date}{a value from 2020-1-22 to today's date}

\item{max.date}{a value from 2020-1-22 to today's date}
}
\value{
a time series plot
}
\description{
Takes the data from updateCovid function and passes it on to plot a time series plot for
every type of data available.
}
\examples{
covidTimeSeries(country = "egypt",min.date = "2020-02-22",max.date = as.Date(today()))
}
\author{
Ali Malek
}
