% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCovid.R
\name{updateCovid}
\alias{updateCovid}
\title{A comprehensive covid-19 dataframe}
\usage{
updateCovid()
}
\value{
A dataframe containing the cumulative daily cases, deaths, and recovers of covid-19 of all countries.
}
\description{
This function gives back a dataframe of total cases, total deaths, and total recoveres from covid-19 in the world.
}
\examples{
updateCovid()
}
\author{
Ali Malek
}
