% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldCovidPlot.R
\name{worldCovidPlot}
\alias{worldCovidPlot}
\title{Plot world map colored with covid-19}
\usage{
worldCovidPlot(day = "2020-09-19", type = "cases")
}
\arguments{
\item{day}{A date value. Default is "2020-09-19"}

\item{type}{type of the data to plot. One of the three values: "cases","deaths", and "recovered".}
}
\value{
a plot
}
\description{
Takes the data from updateCovid function and passes it on to plot a world map colored by the
count of cases of each country.
}
\examples{
worldCovidPlot(day = "2020-09-19", type = "cases")
}
\author{
Ali Malek
}
